+++ src/constants.ts
@@ -664,7 +664,7 @@ export type TreeViewNodeTypes =
 	| 'tracking-status-files'
 	| 'uncommitted-files'
 	| 'workspace-missing-repository'
-	| 'workspaces-view'
+	| 'workspaces'
 	| 'worktree'
 	| 'worktrees';
 
@@ -883,10 +883,13 @@ export type SupportedAIModels =
 	| 'vscode';
 
 export type SecretKeys =
-	| `gitlens.integration.auth:${IntegrationId}|${string}`
-	| `gitlens.integration.auth.cloud:${IntegrationId}|${string}`
+	| IntegrationAuthenticationKeys
 	| `gitlens.${AIProviders}.key`
 	| `gitlens.plus.auth:${Environment}`;
+
+export type IntegrationAuthenticationKeys =
+	| `gitlens.integration.auth:${IntegrationId}|${string}`
+	| `gitlens.integration.auth.cloud:${IntegrationId}|${string}`;
 
 export const enum SyncedStorageKeys {
 	Version = 'gitlens:synced:version',